refactor: code style changes, add interfaces, sort lines, etc.
